Move MultilineParameterInspectionTests for QuickFixes to relevant files under RD.QuickFixes
